# openapi.json
{"openapi":"3.1.0","info":{"title":"Tech Challenge API","description":"API para o Tech Challenge FIAP","contact":{"name":"Onofre Lima"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Usuario","description":"API para gerenciamento de usuários"},{"name":"Autenticação com JWT","description":"API para autenticação de usuários"},{"name":"Autenticação","description":"API para autenticação de usuários"}],"paths":{"/api/v1/usuarios/{id}":{"put":{"tags":["Usuario"],"summary":"Atualizar usuário","description":"Atualiza um usuário existente no sistema","operationId":"atualizarUsuario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuário"}}},"required":true},"responses":{"204":{"description":"Usuário atualizado com sucesso"},"404":{"description":"Usuário não encontrado"},"400":{"description":"Requisição inválida"},"422":{"description":"Erro de validação dos dados enviados"},"500":{"description":"Erro interno do servidor"}}},"delete":{"tags":["Usuario"],"summary":"Deletar usuário","description":"Deleta um usuário existente pelo ID","operationId":"deletarUsuario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}},"required":true},"responses":{"204":{"description":"Usuário deletado com sucesso"},"400":{"description":"Requisição inválida"},"404":{"description":"Usuário não encontrado"}}}},"/api/v1/usuarios/{id}/senha":{"put":{"tags":["Usuario"],"summary":"Atualizar senha do usuário","description":"Atualiza a senha de um usuário existente no sistema","operationId":"atualizarSenha","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Senha"}}},"required":true},"responses":{"204":{"description":"Senha do usuário atualizada com sucesso"},"404":{"description":"Usuário não encontrado"},"400":{"description":"Requisição inválida"},"422":{"description":"Erro de validação dos dados enviados"},"500":{"description":"Erro interno do servidor"}}}},"/api/v1/usuarios":{"get":{"tags":["Usuario"],"summary":"Listar usuários","description":"Retorna uma lista de todos os) usuários cadastrados no sistema","operationId":"listarUsuarios","responses":{"200":{"description":"Lista de usuários retornada com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Usuário Response"}}}}},"204":{"description":"Nenhum usuário encontrado","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Usuário Response"}}}}}}},"post":{"tags":["Usuario"],"summary":"Salvar usuário","description":"Salva um novo usuário no sistema","operationId":"salvar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuário"}}},"required":true},"responses":{"201":{"description":"Usuário criado com sucesso"},"400":{"description":"Requisição inválida"},"422":{"description":"Erro de validação dos dados enviados"},"500":{"description":"Erro interno do servidor"}}}},"/api/v1/auth/login":{"post":{"tags":["Autenticação com JWT"],"summary":"Login de usuário com JWT","description":"Autentica um usuário com base no login e senha fornecidos","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Authentication"}}},"required":true},"responses":{"200":{"description":"Login realizado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"401":{"description":"E-mail ou senha inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/v1/autenticacao/login":{"post":{"tags":["Autenticação"],"summary":"Login de usuário","description":"Autentica um usuário com base no login e senha fornecidos","operationId":"login_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}},"required":true},"responses":{"200":{"description":"Login realizado com sucesso","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"E-mail ou senha inválidos","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/usuarios/{nome}":{"get":{"tags":["Usuario"],"summary":"Buscar usuário por nome","description":"Busca usuários pelo nome fornecido","operationId":"buscarUsuarioPorNome","parameters":[{"name":"nome","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Usuários encontrados com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Usuário Response"}}}}},"404":{"description":"Nenhum usuário encontrado com o nome fornecido","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Usuário Response"}}}}}}}}},"components":{"schemas":{"Endereço":{"type":"object","description":"DTO para representar o endereço do usuário","properties":{"rua":{"type":"string","minLength":1},"numero":{"type":"string","minLength":1},"cidade":{"type":"string","minLength":1},"estado":{"type":"string","minLength":1},"cep":{"type":"string","minLength":1}},"required":["cep","cidade","estado","numero","rua"]},"Usuário":{"type":"object","description":"DTO para representar um usuário","properties":{"nome":{"type":"string","minLength":1},"email":{"type":"string","format":"email"},"login":{"type":"string","minLength":1},"senha":{"type":"string","minLength":1},"endereco":{"$ref":"#/components/schemas/Endereço"},"tipoUsuario":{"type":"integer","format":"int32"}},"required":["login","nome","senha","tipoUsuario"]},"Senha":{"type":"object","description":"DTO para representar alteração de senha do usuário","properties":{"novaSenha":{"type":"string","minLength":1},"confirmaSenha":{"type":"string","minLength":1}},"required":["confirmaSenha","novaSenha"]},"Authentication":{"type":"object","description":"DTO para representar a autenticação do usuário","properties":{"login":{"type":"string","minLength":1},"senha":{"type":"string","minLength":1}},"required":["login","senha"]},"LoginResponse":{"type":"object","description":"DTO para representar o Token JWT de resposta da autenticação do usuário","properties":{"token":{"type":"string"}}},"Login":{"type":"object","description":"DTO para representar o login do usuário","properties":{"login":{"type":"string","minLength":1},"senha":{"type":"string","minLength":1}},"required":["login","senha"]},"Usuário Response":{"type":"object","description":"DTO para representar o retorno de um usuário sem informações sensíveis","properties":{"nome":{"type":"string"},"email":{"type":"string"},"login":{"type":"string"},"endereco":{"$ref":"#/components/schemas/Endereço"},"tipoUsuario":{"type":"integer","format":"int32"}}}}}}